+++ config/items/updater.sh
@@ -3,7 +3,7 @@
 sketchybar --add item updater right \
   --set updater \
     width=46 \
-    update_freq=21600 \
+    update_freq=1800 \
     icon=$ICON_UPDATE \
     icon.color=$PINK \
     drawing=off \